implement SetLoggingParameters() and also GetLevel() interface methods
As the logger implementations are very different and may have very
distinct features which are desirable to use, a generic interface to
activate these features is introduced:

LoggerFactory.SetLoggingParameters(LoggingParameters) error

In this case it is a map[string]interface{}.

It's up to logger adaptor implementation to validate the parameters and
indicate with appropriate error, if required.

Fixes #2.